tidy tidy
Tidy up our tidy.zig!

The main innovation here is introduciton of `Errors` sink for errors,
such that we always report all errors and not just the first one. It
also makes it possible to actually check our tidy checks, some of which
are fairly non-trivial.

Other than that, just moving the code around to make the overal
structure more obvious!